WIP: validate on CRUD endpoints
